+++ config_modules_vmware/controllers/vcenter/vm_migrate_encryption_policy.py
@@ -1,11 +1,13 @@
 # Copyright 2024 Broadcom. All Rights Reserved.
 import logging
+import re
 from typing import Any
 from typing import Dict
 from typing import List
 from typing import Tuple
 
 from pyVmomi import vim  # pylint: disable=E0401
+from pyVmomi import vmodl  # pylint: disable=E0401
 
 from config_modules_vmware.controllers.base_controller import BaseController
 from config_modules_vmware.framework.auth.contexts.base_context import BaseContext
@@ -155,6 +157,94 @@ class VmMigrateEncryptionPolicy(BaseController):
         """
         pass  # pylint: disable=unnecessary-pass
 
+    def _vm_name_check(self, vm_name: str, exclude_vm_names: List) -> bool:
+        """
+        Check if a given vm name matches any name/pattern in exclude list.
+
+        :param vm_name: vm name.
+        :type vm_name: str
+        :param exclude_vm_names: a list of vm names/patterns to compare.
+        :type exclude_vm_names: List
+        :return: True if vm_name matches any name or pattern in exclude list
+        :rtype: bool
+        """
+        for exclude_vm_name in exclude_vm_names:
+            if re.match(exclude_vm_name, vm_name, re.IGNORECASE):
+                logger.debug(f"VM name: {vm_name} -match-: {exclude_vm_name}")
+                return True
+        return False
+
+    def _filter_vm_configs(self, all_vm_migrate_encryption_configs: List, exclude_vms: Dict) -> List:
+        """
+        Remove vm configs in vm exclude list.
+
+        :param all_vm_migrate_encryption_configs: all vm configs.
+        :type all_vm_migrate_encryption_configs: List
+        :param exclude_vms: user input VM exclude patterns.
+        :type exclude_vms: Dict
+        :return: a list of configs to compare with desired input.
+        :rtype: List
+        """
+        all_vm_configs = []
+        excluded_vms = {}
+        vm_type = exclude_vms.get("vm_type", {})
+        exclude_vm_in_bad_state = vm_type.get("vm_disconnected", False)
+        exclude_vm_fully_encrypted = vm_type.get("vm_fully_encrypted", False)
+        exclude_vm_names = exclude_vms.get("vm_name_match", [])
+        for vm_migrate_encryption_config in all_vm_migrate_encryption_configs:
+            vm_name = vm_migrate_encryption_config["vm_name"]
+            vm_state = vm_migrate_encryption_config["vm_state"]
+            migrate_policy = vm_migrate_encryption_config["migrate_encryption_policy"]
+            # check if vm in bad state
+            if vm_migrate_encryption_config["vm_state"] != "connected" and exclude_vm_in_bad_state:
+                logger.debug(
+                    f"Exclude this VM -  name: {vm_name}, state: {vm_state}, policy: {migrate_policy} - bad state"
+                )
+                excluded_vms.setdefault("disconnected", []).append(vm_name)
+                continue
+            # check if vm is fully encrypted
+            if vm_migrate_encryption_config["vm_encrypted"] and exclude_vm_fully_encrypted:
+                logger.debug(
+                    f"Exclude this VM -  name: {vm_name}, state: {vm_state}, policy: {migrate_policy} - encrypted"
+                )
+                excluded_vms.setdefault("encrypted", []).append(vm_name)
+                continue
+            # check if vm name matches the name in exclude list
+            if self._vm_name_check(vm_name, exclude_vm_names):
+                logger.debug(
+                    f"Exclude this VM -  name: {vm_name}, state: {vm_state}, policy: {migrate_policy} - name match"
+                )
+                excluded_vms.setdefault("name_match", []).append(vm_name)
+                continue
+            vm_config = {
+                "vm_name": vm_name,
+                "path": vm_migrate_encryption_config["path"],
+                "migrate_encryption_policy": migrate_policy,
+            }
+            all_vm_configs.append(vm_config)
+
+        logger.info(f"Excluded disconnected VMs: {','.join(excluded_vms.get('disconnected', []))}")
+        logger.info(f"Excluded fully encrypted VMs: {','.join(excluded_vms.get('encrypted', []))}")
+        logger.info(f"Excluded name match VMs: {','.join(excluded_vms.get('name_match', []))}")
+        logger.debug(f"All VM configs: {all_vm_configs}")
+        return all_vm_configs
+
+    def _is_vm_deleted_exception(self, cause) -> bool:
+        """
+        Check if exception received when get config is because VM been deleted.
+
+        :param cause: exception cause.
+        :type cause: Exception
+        :return: True if exception case is "vmodl.fault.ManagedObjectNotFound" otherwise False
+        :rtype: bool
+        """
+
+        while cause:
+            if isinstance(cause, vmodl.fault.ManagedObjectNotFound):
+                return True
+            cause = getattr(cause, "__cause__", None)
+        return False
+
     def __get_all_vm_migrate_encryption_policy(self, vc_vmomi_client: VcVmomiClient) -> List[Dict]:
         """
         Get all VM migrate Encryption policies.
@@ -168,12 +258,26 @@ class VmMigrateEncryptionPolicy(BaseController):
         all_vm_refs = vc_vmomi_client.get_objects_by_vimtype(vim.VirtualMachine)
 
         for vm_ref in all_vm_refs:
-            vm_migrate_encryption_config = {
-                "vm_name": vm_ref.name,
-                "path": vc_vmomi_client.get_vm_path_in_datacenter(vm_ref),
-                "migrate_encryption_policy": vm_ref.config.migrateEncryption,
-            }
-            all_vm_migrate_encryption_configs.append(vm_migrate_encryption_config)
+            try:
+                logger.debug(
+                    f"VM name: {vm_ref.name}, conn state: {vm_ref.runtime.connectionState}, power state: {vm_ref.runtime.powerState}"
+                )
+                vm_migrate_encryption_config = {
+                    "vm_name": vm_ref.name,
+                    "path": vc_vmomi_client.get_vm_path_in_datacenter(vm_ref),
+                    "migrate_encryption_policy": vm_ref.config.migrateEncryption
+                    if vm_ref.config and hasattr(vm_ref.config, "migrateEncryption")
+                    else "None",
+                    "vm_state": vm_ref.runtime.connectionState,
+                    "vm_encrypted": True
+                    if vm_ref.config and hasattr(vm_ref.config, "keyId") and vm_ref.config.keyId is not None
+                    else False,
+                }
+                all_vm_migrate_encryption_configs.append(vm_migrate_encryption_config)
+            except Exception as e:
+                logger.exception(f"Exception when get configs: {e}")
+                if not self._is_vm_deleted_exception(e):
+                    raise
         return all_vm_migrate_encryption_configs
 
     def _get_data_center(self, vm_ref):
@@ -225,8 +329,21 @@ class VmMigrateEncryptionPolicy(BaseController):
             errors.append(f"Resource pool for VM: {vm_ref.name} not found")
         return resource_pool, errors
 
+    def _is_in_non_compiliant_list(self, vm_name: str, non_compliant_configs: List) -> bool:
+        """
+        Check if a given vm name is in non compliant vm list.
+
+        :param vm_name: vm name.
+        :type vm_name: str
+        :param non_compliant_configs: non compliant vm config list.
+        :type non_compliant_configs: List
+        :return: True if vm_name matches any item in non compliant config list
+        :rtype: bool
+        """
+        return any(item.get("vm_name") == vm_name for item in non_compliant_configs)
+
     def __set_vm_migrate_encryption_policy_for_all_non_compliant_vms(
-        self, vc_vmomi_client: VcVmomiClient, desired_values: Dict
+        self, vc_vmomi_client: VcVmomiClient, desired_values: Dict, non_compliant_configs: List
     ) -> Tuple[List[dict], List[dict], List[str]]:
         """
         Set VM migrate Encryption policies for all non-compliant VMs.
@@ -269,60 +386,77 @@ class VmMigrateEncryptionPolicy(BaseController):
         remediated_desired = []
 
         for vm_ref in all_vm_refs:
-            vm_path = vc_vmomi_client.get_vm_path_in_datacenter(vm_ref)
-            current_vm_migrate_encryption_policy = vm_ref.config.migrateEncryption
-            override_vm_migrate_encryption_policy, exclude_flag = next(
-                (
-                    (override.get(DESIRED_KEY), override.get(EXCLUDE_THIS_VM))
-                    for override in overrides
-                    if override[VM_NAME] == vm_ref.name and override[PATH] == vm_path
-                ),
-                (None, None),
-            )
-            desired_vm_migrate_policy = (
-                override_vm_migrate_encryption_policy
-                if override_vm_migrate_encryption_policy is not None
-                else desired_global_vm_migrate_encryption_policy
-            )
-            if current_vm_migrate_encryption_policy != desired_vm_migrate_policy and not exclude_flag:
-                logger.info(f"Setting VM migrate policy {desired_vm_migrate_policy} on VM {vm_ref.name}")
-                # continue to remediate next vm if hitting any errors
-                try:
-                    template = vm_ref.config.template
-                    if template:
-                        # for VM template, convert to VM during remediation
-                        resource_pool, errs = self._get_resource_pool(vm_ref)
-                        if errs:
-                            errors.append(errs)
-                            continue
-                        logger.debug(f"Resource pool for convert template to VM: {resource_pool}")
-                        vm_ref.MarkAsVirtualMachine(pool=resource_pool)
-                        logger.debug(f"Converted VM template to VM, template flag: {vm_ref.config.template}")
-
-                    config_spec = vim.vm.ConfigSpec()
-                    config_spec.migrateEncryption = desired_vm_migrate_policy
-                    task = vm_ref.ReconfigVM_Task(config_spec)
-                    vc_vmomi_client.wait_for_task(task=task)
-
-                    if template:
-                        # for VM template, convert it back to template after remediation
-                        vm_ref.MarkAsTemplate()
-                        logger.debug(f"Converted VM back to VM template, template flag: {vm_ref.config.template}")
-
-                    remediated.append(
-                        {VM_NAME: vm_ref.name, PATH: vm_path, DESIRED_KEY: current_vm_migrate_encryption_policy}
-                    )
-                    remediated_desired.append(
-                        {VM_NAME: vm_ref.name, PATH: vm_path, DESIRED_KEY: desired_vm_migrate_policy}
-                    )
-                except Exception as e:
-                    logger.exception(f"An error occurred: {e}")
-                    errors.append(f"Failed to remediate VM: {vm_ref.name} - {str(e)}")
-            else:
-                logger.info(
-                    f"VM {vm_ref.name} already has desired migrate policy {desired_vm_migrate_policy},"
-                    f" no remediation required. Or exclude flag present - {exclude_flag}"
+            try:
+                vm_name = vm_ref.name
+                if not self._is_in_non_compiliant_list(vm_name, non_compliant_configs):
+                    continue
+                vm_path = vc_vmomi_client.get_vm_path_in_datacenter(vm_ref)
+                current_vm_migrate_encryption_policy = (
+                    vm_ref.config.migrateEncryption
+                    if vm_ref.config and hasattr(vm_ref.config, "migrateEncryption")
+                    else "None"
                 )
+                override_vm_migrate_encryption_policy, exclude_flag = next(
+                    (
+                        (override.get(DESIRED_KEY), override.get(EXCLUDE_THIS_VM))
+                        for override in overrides
+                        if override[VM_NAME] == vm_name and override[PATH] == vm_path
+                    ),
+                    (None, None),
+                )
+                desired_vm_migrate_policy = (
+                    override_vm_migrate_encryption_policy
+                    if override_vm_migrate_encryption_policy is not None
+                    else desired_global_vm_migrate_encryption_policy
+                )
+                if current_vm_migrate_encryption_policy != desired_vm_migrate_policy and not exclude_flag:
+                    logger.info(f"Setting VM migrate policy {desired_vm_migrate_policy} on VM {vm_name}")
+                    # continue to remediate next vm if hitting any errors
+                    try:
+                        template = vm_ref.config.template
+                        if template:
+                            # for VM template, convert to VM during remediation
+                            resource_pool, errs = self._get_resource_pool(vm_ref)
+                            if errs:
+                                errors.append(errs)
+                                continue
+                            logger.debug(f"Resource pool for convert template to VM: {resource_pool}")
+                            vm_ref.MarkAsVirtualMachine(pool=resource_pool)
+                            logger.debug(f"Converted VM template to VM, template flag: {vm_ref.config.template}")
+
+                        config_spec = vim.vm.ConfigSpec()
+                        config_spec.migrateEncryption = desired_vm_migrate_policy
+                        task = vm_ref.ReconfigVM_Task(config_spec)
+                        vc_vmomi_client.wait_for_task(task=task, timeout=120)
+
+                        if template:
+                            # for VM template, convert it back to template after remediation
+                            vm_ref.MarkAsTemplate()
+                            logger.debug(f"Converted VM back to VM template, template flag: {vm_ref.config.template}")
+
+                        remediated.append(
+                            {VM_NAME: vm_name, PATH: vm_path, DESIRED_KEY: current_vm_migrate_encryption_policy}
+                        )
+                        remediated_desired.append(
+                            {VM_NAME: vm_name, PATH: vm_path, DESIRED_KEY: desired_vm_migrate_policy}
+                        )
+                    except Exception as e:
+                        logger.exception(f"An error occurred: {e}")
+                        if not self._is_vm_deleted_exception(e):
+                            errors.append(f"Failed to remediate VM: {vm_name} - {str(e)}")
+                else:
+                    if exclude_flag:
+                        logger.info(
+                            f"VM {vm_name} is excluded from remediation,  exclude flag present - {exclude_flag}"
+                        )
+                    else:
+                        logger.info(
+                            f"VM {vm_name} already has desired migrate policy {desired_vm_migrate_policy}, no remediation required."
+                        )
+            except Exception as e:
+                logger.exception(f"An error occurred: {e}")
+                if not self._is_vm_deleted_exception(e):
+                    errors.append(f"Failed to remediate VM: {vm_name} - {str(e)}")
 
         return remediated, remediated_desired, errors
 
@@ -394,6 +528,11 @@ class VmMigrateEncryptionPolicy(BaseController):
         if errors:
             return {consts.STATUS: ComplianceStatus.FAILED, consts.ERRORS: errors}
 
+        # check if need to exclude any VMs.
+        exclude_vms = desired_values.get("exclude_vms", {})
+        logger.debug(f"User input VM patterns to be excluded from compliance check: {exclude_vms}")
+        all_vm_migrate_encryption_configs = self._filter_vm_configs(all_vm_migrate_encryption_configs, exclude_vms)
+
         non_compliant_configs, desired_configs = self.__get_non_compliant_configs(
             all_vm_migrate_encryption_configs, desired_values
         )
@@ -450,9 +589,10 @@ class VmMigrateEncryptionPolicy(BaseController):
             errors = result[consts.ERRORS]
             return {consts.STATUS: RemediateStatus.FAILED, consts.ERRORS: errors}
 
+        non_compliant_configs = result.get(consts.CURRENT, [])
         vc_vmomi_client = context.vc_vmomi_client()
         remediated, remediated_desired, errors = self.__set_vm_migrate_encryption_policy_for_all_non_compliant_vms(
-            vc_vmomi_client, desired_values
+            vc_vmomi_client, desired_values, non_compliant_configs
         )
 
         if not errors: